More work on the various scenarios, this will allow to understand the neural networks way better and make progress on the speed of the IACars

+++ src/main/java/com/peacefulotter/selfdrivingcar/game/GameLoop.java
@@ -47,7 +47,7 @@ public class GameLoop extends AnimationTimer
         {
             unprocessedTime -= FRAME_TIME;
             render = true;
-            circuit.update( (float) FRAME_TIME );
+            circuit.update( FRAME_TIME );
 
             if (framesCounter >= Time.SECOND)
             {
